titan-voice-agent/requirements.txt
livekit-agents>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-plugins-turn-detector>=1.0.0
openai>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
